Rename misc module to util since it's more clear
